requirements.txt
pypng==0.20220715.0
Pillow==10.0.0
pygame==2.5.1